+++ src/strings.h
@@ -26,7 +26,7 @@ extern void str_append_uint32(String *this, uint32_t integer);
 extern void str_append_int64_t(String *this, int64_t integer);
 extern void str_append_double(String *this, double dbl);
 extern void str_free(String *this);
-extern int str_size(String *this);
+extern size_t str_size(String *this);
 extern char* str_data(String *this);
 extern char str_char_at(String *this, int position);
 extern void str_appendCs_size(String *this, const char *that, uint32_t size);
